Getting our basic command (get, make, install) working in Drupal 8.
